+++ tests/framework/db/ActiveRecordTest.php
@@ -32,7 +32,6 @@ use yiiunit\data\ar\Profile;
 use yiiunit\data\ar\ProfileWithConstructor;
 use yiiunit\data\ar\Type;
 use yiiunit\framework\ar\ActiveRecordTestTrait;
-use yiiunit\framework\db\cubrid\ActiveRecordTest as CubridActiveRecordTest;
 use yiiunit\TestCase;
 
 abstract class ActiveRecordTest extends DatabaseTestCase
@@ -1166,13 +1165,6 @@ abstract class ActiveRecordTest extends DatabaseTestCase
         $this->assertEquals(1.23, $model->float_col2);
         $this->assertEquals(33.22, $model->numeric_col);
         $this->assertEquals(true, $model->bool_col2);
-
-        if ($this instanceof CubridActiveRecordTest) {
-            // cubrid has non-standard timestamp representation
-            $this->assertEquals('12:00:00 AM 01/01/2002', $model->time);
-        } else {
-            $this->assertEquals('2002-01-01 00:00:00', $model->time);
-        }
 
         $model = new Type();
         $model->char_col2 = 'not something';