+++ tests/framework/db/CommandTest.php
@@ -636,7 +636,6 @@ SQL;
             case 'pgsql':
                 $expression = "EXTRACT(YEAR FROM TIMESTAMP 'now')";
             break;
-            case 'cubrid':
             case 'mysql':
                 $expression = 'YEAR(NOW())';
             break;