+++ tests/framework/db/DatabaseTestCase.php
@@ -118,7 +118,6 @@ abstract class DatabaseTestCase extends TestCase
             case 'mysql':
             case 'sqlite':
                 return str_replace(['[[', ']]'], '`', $sql);
-            case 'cubrid':
             case 'oci':
                 return str_replace(['[[', ']]'], '"', $sql);
             case 'pgsql':