+++ tests/framework/db/QueryBuilderTest.php
@@ -8,7 +8,6 @@
 namespace yiiunit\framework\db;
 
 use yii\db\conditions\BetweenColumnsCondition;
-use yii\db\cubrid\QueryBuilder as CubridQueryBuilder;
 use yii\db\Expression;
 use yii\db\mysql\QueryBuilder as MysqlQueryBuilder;
 use yii\db\pgsql\QueryBuilder as PgsqlQueryBuilder;
@@ -57,8 +56,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 return new SqliteQueryBuilder($connection);
             case 'pgsql':
                 return new PgsqlQueryBuilder($connection);
-            case 'cubrid':
-                return new CubridQueryBuilder($connection);
         }
         throw new \Exception('Test is not implemented for ' . $this->driverName);
     }
@@ -77,9 +74,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'bigint(20)',
                     'postgres' => 'bigint',
                     'sqlite' => 'bigint',
-                    'oci' => 'NUMBER(20)',
-                    'sqlsrv' => 'bigint',
-                    'cubrid' => 'bigint',
                 ],
             ],
             [
@@ -89,9 +83,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'bigint(20) NOT NULL',
                     'postgres' => 'bigint NOT NULL',
                     'sqlite' => 'bigint NOT NULL',
-                    'oci' => 'NUMBER(20) NOT NULL',
-                    'sqlsrv' => 'bigint NOT NULL',
-                    'cubrid' => 'bigint NOT NULL',
                 ],
             ],
             [
@@ -101,9 +92,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'bigint(20) CHECK (value > 5)',
                     'postgres' => 'bigint CHECK (value > 5)',
                     'sqlite' => 'bigint CHECK (value > 5)',
-                    'oci' => 'NUMBER(20) CHECK (value > 5)',
-                    'sqlsrv' => 'bigint CHECK (value > 5)',
-                    'cubrid' => 'bigint CHECK (value > 5)',
                 ],
             ],
             [
@@ -113,9 +101,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'bigint(8)',
                     'postgres' => 'bigint',
                     'sqlite' => 'bigint',
-                    'oci' => 'NUMBER(8)',
-                    'sqlsrv' => 'bigint',
-                    'cubrid' => 'bigint',
                 ],
             ],
             [
@@ -125,9 +110,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'bigint(8) CHECK (value > 5)',
                     'postgres' => 'bigint CHECK (value > 5)',
                     'sqlite' => 'bigint CHECK (value > 5)',
-                    'oci' => 'NUMBER(8) CHECK (value > 5)',
-                    'sqlsrv' => 'bigint CHECK (value > 5)',
-                    'cubrid' => 'bigint CHECK (value > 5)',
                 ],
             ],
             [
@@ -146,9 +128,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'blob',
                     'postgres' => 'bytea',
                     'sqlite' => 'blob',
-                    'oci' => 'BLOB',
-                    'sqlsrv' => 'varbinary(max)',
-                    'cubrid' => 'blob',
                 ],
             ],
             [
@@ -157,8 +136,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'tinyint(1) NOT NULL DEFAULT 1',
                     'sqlite' => 'boolean NOT NULL DEFAULT 1',
-                    'sqlsrv' => 'tinyint(1) NOT NULL DEFAULT 1',
-                    'cubrid' => 'smallint NOT NULL DEFAULT 1',
                 ],
             ],
             [
@@ -168,9 +145,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'tinyint(1)',
                     'postgres' => 'boolean',
                     'sqlite' => 'boolean',
-                    'oci' => 'NUMBER(1)',
-                    'sqlsrv' => 'tinyint(1)',
-                    'cubrid' => 'smallint',
                 ],
             ],
             [
@@ -179,7 +153,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'char(1) CHECK (value LIKE "test%")',
                     'sqlite' => 'char(1) CHECK (value LIKE "test%")',
-                    'cubrid' => 'char(1) CHECK (value LIKE "test%")',
                 ],
             ],
             [
@@ -189,8 +162,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'char(1) NOT NULL',
                     'postgres' => 'char(1) NOT NULL',
                     'sqlite' => 'char(1) NOT NULL',
-                    'oci' => 'CHAR(1) NOT NULL',
-                    'cubrid' => 'char(1) NOT NULL',
                 ],
             ],
             [
@@ -199,7 +170,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'char(6) CHECK (value LIKE "test%")',
                     'sqlite' => 'char(6) CHECK (value LIKE "test%")',
-                    'cubrid' => 'char(6) CHECK (value LIKE "test%")',
                 ],
             ],
             [
@@ -209,8 +179,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'char(6)',
                     'postgres' => 'char(6)',
                     'sqlite' => 'char(6)',
-                    'oci' => 'CHAR(6)',
-                    'cubrid' => 'char(6)',
                 ],
             ],
             [
@@ -220,8 +188,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'char(1)',
                     'postgres' => 'char(1)',
                     'sqlite' => 'char(1)',
-                    'oci' => 'CHAR(1)',
-                    'cubrid' => 'char(1)',
                 ],
             ],
             //[
@@ -231,8 +197,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
             //        'mysql' => ,
             //        'postgres' => ,
             //        'sqlite' => ,
-            //        'sqlsrv' => ,
-            //        'cubrid' => ,
             //    ],
             //],
             [
@@ -242,9 +206,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'date NOT NULL',
                     'postgres' => 'date NOT NULL',
                     'sqlite' => 'date NOT NULL',
-                    'oci' => 'DATE NOT NULL',
-                    'sqlsrv' => 'date NOT NULL',
-                    'cubrid' => 'date NOT NULL',
                 ],
             ],
             [
@@ -254,9 +215,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'date',
                     'postgres' => 'date',
                     'sqlite' => 'date',
-                    'oci' => 'DATE',
-                    'sqlsrv' => 'date',
-                    'cubrid' => 'date',
                 ],
             ],
             //[
@@ -266,8 +224,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
             //        'mysql' => ,
             //        'postgres' => ,
             //        'sqlite' => ,
-            //        'sqlsrv' => ,
-            //        'cubrid' => ,
             //    ],
             //],
             [
@@ -277,9 +233,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'datetime NOT NULL',
                     'postgres' => 'timestamp(0) NOT NULL',
                     'sqlite' => 'datetime NOT NULL',
-                    'oci' => 'TIMESTAMP NOT NULL',
-                    'sqlsrv' => 'datetime NOT NULL',
-                    'cubrid' => 'datetime NOT NULL',
                 ],
             ],
             [
@@ -289,9 +242,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'datetime',
                     'postgres' => 'timestamp(0)',
                     'sqlite' => 'datetime',
-                    'oci' => 'TIMESTAMP',
-                    'sqlsrv' => 'datetime',
-                    'cubrid' => 'datetime',
                 ],
             ],
             [
@@ -301,9 +251,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(10,0) CHECK (value > 5.6)',
                     'postgres' => 'numeric(10,0) CHECK (value > 5.6)',
                     'sqlite' => 'decimal(10,0) CHECK (value > 5.6)',
-                    'oci' => 'NUMBER CHECK (value > 5.6)',
-                    'sqlsrv' => 'decimal(18,0) CHECK (value > 5.6)',
-                    'cubrid' => 'decimal(10,0) CHECK (value > 5.6)',
                 ],
             ],
             [
@@ -313,9 +260,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(10,0) NOT NULL',
                     'postgres' => 'numeric(10,0) NOT NULL',
                     'sqlite' => 'decimal(10,0) NOT NULL',
-                    'oci' => 'NUMBER NOT NULL',
-                    'sqlsrv' => 'decimal(18,0) NOT NULL',
-                    'cubrid' => 'decimal(10,0) NOT NULL',
                 ],
             ],
             [
@@ -325,9 +269,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(12,4) CHECK (value > 5.6)',
                     'postgres' => 'numeric(12,4) CHECK (value > 5.6)',
                     'sqlite' => 'decimal(12,4) CHECK (value > 5.6)',
-                    'oci' => 'NUMBER CHECK (value > 5.6)',
-                    'sqlsrv' => 'decimal(12,4) CHECK (value > 5.6)',
-                    'cubrid' => 'decimal(12,4) CHECK (value > 5.6)',
                 ],
             ],
             [
@@ -337,9 +278,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(12,4)',
                     'postgres' => 'numeric(12,4)',
                     'sqlite' => 'decimal(12,4)',
-                    'oci' => 'NUMBER',
-                    'sqlsrv' => 'decimal(12,4)',
-                    'cubrid' => 'decimal(12,4)',
                 ],
             ],
             [
@@ -349,9 +287,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(10,0)',
                     'postgres' => 'numeric(10,0)',
                     'sqlite' => 'decimal(10,0)',
-                    'oci' => 'NUMBER',
-                    'sqlsrv' => 'decimal(18,0)',
-                    'cubrid' => 'decimal(10,0)',
                 ],
             ],
             [
@@ -361,9 +296,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'double CHECK (value > 5.6)',
                     'postgres' => 'double precision CHECK (value > 5.6)',
                     'sqlite' => 'double CHECK (value > 5.6)',
-                    'oci' => 'NUMBER CHECK (value > 5.6)',
-                    'sqlsrv' => 'float CHECK (value > 5.6)',
-                    'cubrid' => 'double(15) CHECK (value > 5.6)',
                 ],
             ],
             [
@@ -373,9 +305,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'double NOT NULL',
                     'postgres' => 'double precision NOT NULL',
                     'sqlite' => 'double NOT NULL',
-                    'oci' => 'NUMBER NOT NULL',
-                    'sqlsrv' => 'float NOT NULL',
-                    'cubrid' => 'double(15) NOT NULL',
                 ],
             ],
             [
@@ -385,9 +314,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'double CHECK (value > 5.6)',
                     'postgres' => 'double precision CHECK (value > 5.6)',
                     'sqlite' => 'double CHECK (value > 5.6)',
-                    'oci' => 'NUMBER CHECK (value > 5.6)',
-                    'sqlsrv' => 'float CHECK (value > 5.6)',
-                    'cubrid' => 'double(16) CHECK (value > 5.6)',
                 ],
             ],
             [
@@ -396,9 +322,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'double',
                     'sqlite' => 'double',
-                    'oci' => 'NUMBER',
-                    'sqlsrv' => 'float',
-                    'cubrid' => 'double(16)',
                 ],
             ],
             [
@@ -408,9 +331,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'double',
                     'postgres' => 'double precision',
                     'sqlite' => 'double',
-                    'oci' => 'NUMBER',
-                    'sqlsrv' => 'float',
-                    'cubrid' => 'double(15)',
                 ],
             ],
             [
@@ -420,9 +340,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'float CHECK (value > 5.6)',
                     'postgres' => 'double precision CHECK (value > 5.6)',
                     'sqlite' => 'float CHECK (value > 5.6)',
-                    'oci' => 'NUMBER CHECK (value > 5.6)',
-                    'sqlsrv' => 'float CHECK (value > 5.6)',
-                    'cubrid' => 'float(7) CHECK (value > 5.6)',
                 ],
             ],
             [
@@ -432,9 +349,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'float NOT NULL',
                     'postgres' => 'double precision NOT NULL',
                     'sqlite' => 'float NOT NULL',
-                    'oci' => 'NUMBER NOT NULL',
-                    'sqlsrv' => 'float NOT NULL',
-                    'cubrid' => 'float(7) NOT NULL',
                 ],
             ],
             [
@@ -444,9 +358,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'float CHECK (value > 5.6)',
                     'postgres' => 'double precision CHECK (value > 5.6)',
                     'sqlite' => 'float CHECK (value > 5.6)',
-                    'oci' => 'NUMBER CHECK (value > 5.6)',
-                    'sqlsrv' => 'float CHECK (value > 5.6)',
-                    'cubrid' => 'float(16) CHECK (value > 5.6)',
                 ],
             ],
             [
@@ -455,9 +366,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'float',
                     'sqlite' => 'float',
-                    'oci' => 'NUMBER',
-                    'sqlsrv' => 'float',
-                    'cubrid' => 'float(16)',
                 ],
             ],
             [
@@ -467,9 +375,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'float',
                     'postgres' => 'double precision',
                     'sqlite' => 'float',
-                    'oci' => 'NUMBER',
-                    'sqlsrv' => 'float',
-                    'cubrid' => 'float(7)',
                 ],
             ],
             [
@@ -479,9 +384,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'int(11) CHECK (value > 5)',
                     'postgres' => 'integer CHECK (value > 5)',
                     'sqlite' => 'integer CHECK (value > 5)',
-                    'oci' => 'NUMBER(10) CHECK (value > 5)',
-                    'sqlsrv' => 'int CHECK (value > 5)',
-                    'cubrid' => 'int CHECK (value > 5)',
                 ],
             ],
             [
@@ -491,9 +393,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'int(11) NOT NULL',
                     'postgres' => 'integer NOT NULL',
                     'sqlite' => 'integer NOT NULL',
-                    'oci' => 'NUMBER(10) NOT NULL',
-                    'sqlsrv' => 'int NOT NULL',
-                    'cubrid' => 'int NOT NULL',
                 ],
             ],
             [
@@ -503,9 +402,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'int(8) CHECK (value > 5)',
                     'postgres' => 'integer CHECK (value > 5)',
                     'sqlite' => 'integer CHECK (value > 5)',
-                    'oci' => 'NUMBER(8) CHECK (value > 5)',
-                    'sqlsrv' => 'int CHECK (value > 5)',
-                    'cubrid' => 'int CHECK (value > 5)',
                 ],
             ],
             [
@@ -515,9 +411,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'int(8)',
                     'postgres' => 'integer',
                     'sqlite' => 'integer',
-                    'oci' => 'NUMBER(8)',
-                    'sqlsrv' => 'int',
-                    'cubrid' => 'int',
                 ],
             ],
             [
@@ -527,9 +420,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'int(11)',
                     'postgres' => 'integer',
                     'sqlite' => 'integer',
-                    'oci' => 'NUMBER(10)',
-                    'sqlsrv' => 'int',
-                    'cubrid' => 'int',
                 ],
             ],
             [
@@ -539,9 +429,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(19,4) CHECK (value > 0.0)',
                     'postgres' => 'numeric(19,4) CHECK (value > 0.0)',
                     'sqlite' => 'decimal(19,4) CHECK (value > 0.0)',
-                    'oci' => 'NUMBER(19,4) CHECK (value > 0.0)',
-                    'sqlsrv' => 'decimal(19,4) CHECK (value > 0.0)',
-                    'cubrid' => 'decimal(19,4) CHECK (value > 0.0)',
                 ],
             ],
             [
@@ -551,9 +438,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(19,4) NOT NULL',
                     'postgres' => 'numeric(19,4) NOT NULL',
                     'sqlite' => 'decimal(19,4) NOT NULL',
-                    'oci' => 'NUMBER(19,4) NOT NULL',
-                    'sqlsrv' => 'decimal(19,4) NOT NULL',
-                    'cubrid' => 'decimal(19,4) NOT NULL',
                 ],
             ],
             [
@@ -563,9 +447,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(16,2) CHECK (value > 0.0)',
                     'postgres' => 'numeric(16,2) CHECK (value > 0.0)',
                     'sqlite' => 'decimal(16,2) CHECK (value > 0.0)',
-                    'oci' => 'NUMBER(16,2) CHECK (value > 0.0)',
-                    'sqlsrv' => 'decimal(16,2) CHECK (value > 0.0)',
-                    'cubrid' => 'decimal(16,2) CHECK (value > 0.0)',
                 ],
             ],
             [
@@ -575,9 +456,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(16,2)',
                     'postgres' => 'numeric(16,2)',
                     'sqlite' => 'decimal(16,2)',
-                    'oci' => 'NUMBER(16,2)',
-                    'sqlsrv' => 'decimal(16,2)',
-                    'cubrid' => 'decimal(16,2)',
                 ],
             ],
             [
@@ -587,9 +465,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'decimal(19,4)',
                     'postgres' => 'numeric(19,4)',
                     'sqlite' => 'decimal(19,4)',
-                    'oci' => 'NUMBER(19,4)',
-                    'sqlsrv' => 'decimal(19,4)',
-                    'cubrid' => 'decimal(19,4)',
                 ],
             ],
             [
@@ -599,9 +474,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY CHECK (value > 5)',
                     'postgres' => 'serial NOT NULL PRIMARY KEY CHECK (value > 5)',
                     'sqlite' => 'integer PRIMARY KEY AUTOINCREMENT NOT NULL CHECK (value > 5)',
-                    'oci' => 'NUMBER(10) NOT NULL PRIMARY KEY CHECK (value > 5)',
-                    'sqlsrv' => 'int IDENTITY PRIMARY KEY CHECK (value > 5)',
-                    'cubrid' => 'int NOT NULL AUTO_INCREMENT PRIMARY KEY CHECK (value > 5)',
                 ],
             ],
             [
@@ -609,7 +481,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 $this->primaryKey(8)->check('value > 5'),
                 [
                     'mysql' => 'int(8) NOT NULL AUTO_INCREMENT PRIMARY KEY CHECK (value > 5)',
-                    'oci' => 'NUMBER(8) NOT NULL PRIMARY KEY CHECK (value > 5)',
                 ],
             ],
             [
@@ -617,7 +488,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 $this->primaryKey(8),
                 [
                     'mysql' => 'int(8) NOT NULL AUTO_INCREMENT PRIMARY KEY',
-                    'oci' => 'NUMBER(8) NOT NULL PRIMARY KEY',
                 ],
             ],
             [
@@ -627,9 +497,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY',
                     'postgres' => 'serial NOT NULL PRIMARY KEY',
                     'sqlite' => 'integer PRIMARY KEY AUTOINCREMENT NOT NULL',
-                    'oci' => 'NUMBER(10) NOT NULL PRIMARY KEY',
-                    'sqlsrv' => 'int IDENTITY PRIMARY KEY',
-                    'cubrid' => 'int NOT NULL AUTO_INCREMENT PRIMARY KEY',
                 ],
             ],
             [
@@ -639,9 +506,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'tinyint(2)',
                     'postgres' => 'smallint',
                     'sqlite' => 'tinyint',
-                    'oci' => 'NUMBER(2)',
-                    'sqlsrv' => 'tinyint',
-                    'cubrid' => 'smallint',
                 ],
             ],
             [
@@ -651,7 +515,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'tinyint(3) UNSIGNED',
                     'postgres' => 'smallint UNSIGNED',
                     'sqlite' => 'tinyint UNSIGNED',
-                    'cubrid' => 'smallint UNSIGNED',
                 ],
             ],
             [
@@ -661,9 +524,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'tinyint(3)',
                     'postgres' => 'smallint',
                     'sqlite' => 'tinyint',
-                    'oci' => 'NUMBER(3)',
-                    'sqlsrv' => 'smallint',
-                    'cubrid' => 'smallint',
                 ],
             ],
             [
@@ -673,9 +533,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'smallint(8)',
                     'postgres' => 'smallint',
                     'sqlite' => 'smallint',
-                    'oci' => 'NUMBER(8)',
-                    'sqlsrv' => 'smallint',
-                    'cubrid' => 'smallint',
                 ],
             ],
             [
@@ -685,9 +542,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'smallint(6)',
                     'postgres' => 'smallint',
                     'sqlite' => 'smallint',
-                    'oci' => 'NUMBER(5)',
-                    'sqlsrv' => 'smallint',
-                    'cubrid' => 'smallint',
                 ],
             ],
             [
@@ -696,8 +550,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'varchar(255) CHECK (value LIKE "test%")',
                     'sqlite' => 'varchar(255) CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'varchar(255) CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar(255) CHECK (value LIKE "test%")',
                 ],
             ],
             [
@@ -705,7 +557,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 $this->string()->check('value LIKE \'test%\''),
                 [
                     'postgres' => 'varchar(255) CHECK (value LIKE \'test%\')',
-                    'oci' => 'VARCHAR2(255) CHECK (value LIKE \'test%\')',
                 ],
             ],
             [
@@ -715,9 +566,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'varchar(255) NOT NULL',
                     'postgres' => 'varchar(255) NOT NULL',
                     'sqlite' => 'varchar(255) NOT NULL',
-                    'oci' => 'VARCHAR2(255) NOT NULL',
-                    'sqlsrv' => 'varchar(255) NOT NULL',
-                    'cubrid' => 'varchar(255) NOT NULL',
                 ],
             ],
             [
@@ -726,8 +574,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'varchar(32) CHECK (value LIKE "test%")',
                     'sqlite' => 'varchar(32) CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'varchar(32) CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar(32) CHECK (value LIKE "test%")',
                 ],
             ],
             [
@@ -735,7 +581,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 $this->string(32)->check('value LIKE \'test%\''),
                 [
                     'postgres' => 'varchar(32) CHECK (value LIKE \'test%\')',
-                    'oci' => 'VARCHAR2(32) CHECK (value LIKE \'test%\')',
                 ],
             ],
             [
@@ -745,9 +590,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'varchar(32)',
                     'postgres' => 'varchar(32)',
                     'sqlite' => 'varchar(32)',
-                    'oci' => 'VARCHAR2(32)',
-                    'sqlsrv' => 'varchar(32)',
-                    'cubrid' => 'varchar(32)',
                 ],
             ],
             [
@@ -757,9 +599,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'varchar(255)',
                     'postgres' => 'varchar(255)',
                     'sqlite' => 'varchar(255)',
-                    'oci' => 'VARCHAR2(255)',
-                    'sqlsrv' => 'varchar(255)',
-                    'cubrid' => 'varchar(255)',
                 ],
             ],
             [
@@ -768,8 +607,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'text CHECK (value LIKE "test%")',
                     'sqlite' => 'text CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'text CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar CHECK (value LIKE "test%")',
                 ],
             ],
             [
@@ -777,7 +614,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 $this->text()->check('value LIKE \'test%\''),
                 [
                     'postgres' => 'text CHECK (value LIKE \'test%\')',
-                    'oci' => 'CLOB CHECK (value LIKE \'test%\')',
                 ],
             ],
             [
@@ -787,9 +623,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'text NOT NULL',
                     'postgres' => 'text NOT NULL',
                     'sqlite' => 'text NOT NULL',
-                    'oci' => 'CLOB NOT NULL',
-                    'sqlsrv' => 'text NOT NULL',
-                    'cubrid' => 'varchar NOT NULL',
                 ],
             ],
             [
@@ -798,8 +631,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'text CHECK (value LIKE "test%")',
                     'sqlite' => 'text CHECK (value LIKE "test%")',
-                    'sqlsrv' => 'text CHECK (value LIKE "test%")',
-                    'cubrid' => 'varchar CHECK (value LIKE "test%")',
                 ],
                 Schema::TYPE_TEXT . ' CHECK (value LIKE "test%")',
             ],
@@ -808,7 +639,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 $this->text(255)->check('value LIKE \'test%\''),
                 [
                     'postgres' => 'text CHECK (value LIKE \'test%\')',
-                    'oci' => 'CLOB CHECK (value LIKE \'test%\')',
                 ],
                 Schema::TYPE_TEXT . ' CHECK (value LIKE \'test%\')',
             ],
@@ -819,9 +649,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'text NOT NULL',
                     'postgres' => 'text NOT NULL',
                     'sqlite' => 'text NOT NULL',
-                    'oci' => 'CLOB NOT NULL',
-                    'sqlsrv' => 'text NOT NULL',
-                    'cubrid' => 'varchar NOT NULL',
                 ],
                 Schema::TYPE_TEXT . ' NOT NULL',
             ],
@@ -832,9 +659,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'text',
                     'postgres' => 'text',
                     'sqlite' => 'text',
-                    'oci' => 'CLOB',
-                    'sqlsrv' => 'text',
-                    'cubrid' => 'varchar',
                 ],
                 Schema::TYPE_TEXT,
             ],
@@ -845,9 +669,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'text',
                     'postgres' => 'text',
                     'sqlite' => 'text',
-                    'oci' => 'CLOB',
-                    'sqlsrv' => 'text',
-                    'cubrid' => 'varchar',
                 ],
             ],
             //[
@@ -857,8 +678,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
             //        'mysql' => ,
             //        'postgres' => ,
             //        'sqlite' => ,
-            //        'sqlsrv' => ,
-            //        'cubrid' => ,
             //    ],
             //],
             [
@@ -868,9 +687,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'time NOT NULL',
                     'postgres' => 'time(0) NOT NULL',
                     'sqlite' => 'time NOT NULL',
-                    'oci' => 'TIMESTAMP NOT NULL',
-                    'sqlsrv' => 'time NOT NULL',
-                    'cubrid' => 'time NOT NULL',
                 ],
             ],
             [
@@ -880,9 +696,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'time',
                     'postgres' => 'time(0)',
                     'sqlite' => 'time',
-                    'oci' => 'TIMESTAMP',
-                    'sqlsrv' => 'time',
-                    'cubrid' => 'time',
                 ],
             ],
             //[
@@ -892,8 +705,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
             //        'mysql' => ,
             //        'postgres' => ,
             //        'sqlite' => ,
-            //        'sqlsrv' => ,
-            //        'cubrid' => ,
             //    ],
             //],
             [
@@ -903,9 +714,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'timestamp NOT NULL',
                     'postgres' => 'timestamp(0) NOT NULL',
                     'sqlite' => 'timestamp NOT NULL',
-                    'oci' => 'TIMESTAMP NOT NULL',
-                    'sqlsrv' => 'timestamp NOT NULL',
-                    'cubrid' => 'timestamp NOT NULL',
                 ],
             ],
             [
@@ -915,9 +723,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'timestamp',
                     'postgres' => 'timestamp(0)',
                     'sqlite' => 'timestamp',
-                    'oci' => 'TIMESTAMP',
-                    'sqlsrv' => 'timestamp',
-                    'cubrid' => 'timestamp',
                 ],
             ],
             [
@@ -927,8 +732,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                     'mysql' => 'timestamp NULL DEFAULT NULL',
                     'postgres' => 'timestamp(0) NULL DEFAULT NULL',
                     'sqlite' => 'timestamp NULL DEFAULT NULL',
-                    'sqlsrv' => 'timestamp NULL DEFAULT NULL',
-                    'cubrid' => 'timestamp NULL DEFAULT NULL',
                 ],
             ],
             [
@@ -955,8 +758,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => "int(11) COMMENT 'test comment'",
                     'postgres' => 'integer',
-                    'sqlsrv' => 'int',
-                    'cubrid' => "int COMMENT 'test comment'",
                 ],
             ],
             [
@@ -965,8 +766,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => "int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY COMMENT 'test comment'",
                     'postgres' => 'serial NOT NULL PRIMARY KEY',
-                    'sqlsrv' => 'int IDENTITY PRIMARY KEY',
-                    'cubrid' => "int NOT NULL AUTO_INCREMENT PRIMARY KEY COMMENT 'test comment'",
                 ],
             ],
             [
@@ -975,9 +774,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY FIRST',
                     'postgres' => 'serial NOT NULL PRIMARY KEY',
-                    'oci' => 'NUMBER(10) NOT NULL PRIMARY KEY',
-                    'sqlsrv' => 'int IDENTITY PRIMARY KEY',
-                    'cubrid' => 'int NOT NULL AUTO_INCREMENT PRIMARY KEY FIRST',
                 ],
             ],
             [
@@ -986,9 +782,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'int(11) FIRST',
                     'postgres' => 'integer',
-                    'oci' => 'NUMBER(10)',
-                    'sqlsrv' => 'int',
-                    'cubrid' => 'int FIRST',
                 ],
             ],
             [
@@ -997,9 +790,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'varchar(255) FIRST',
                     'postgres' => 'varchar(255)',
-                    'oci' => 'VARCHAR2(255)',
-                    'sqlsrv' => 'varchar(255)',
-                    'cubrid' => 'varchar(255) FIRST',
                 ],
             ],
             [
@@ -1008,9 +798,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'int(11) NOT NULL FIRST',
                     'postgres' => 'integer NOT NULL',
-                    'oci' => 'NUMBER(10) NOT NULL',
-                    'sqlsrv' => 'int NOT NULL',
-                    'cubrid' => 'int NOT NULL FIRST',
                 ],
             ],
             [
@@ -1019,9 +806,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
                 [
                     'mysql' => 'varchar(255) NOT NULL FIRST',
                     'postgres' => 'varchar(255) NOT NULL',
-                    'oci' => 'VARCHAR2(255) NOT NULL',
-                    'sqlsrv' => 'varchar(255) NOT NULL',
-                    'cubrid' => 'varchar(255) NOT NULL FIRST',
                 ],
             ],
         ];
@@ -1168,7 +952,6 @@ abstract class QueryBuilderTest extends DatabaseTestCase
             [new Expression('NOT (any_expression(:a))', [':a' => 1]), 'NOT (any_expression(:a))', [':a' => 1]],
         ];
         switch ($this->driverName) {
-            case 'sqlsrv':
             case 'sqlite':
                 $conditions = array_merge($conditions, [
                     [['in', ['id', 'name'], [['id' => 1, 'name' => 'foo'], ['id' => 2, 'name' => 'bar']]], '(([[id]] = :qp0 AND [[name]] = :qp1) OR ([[id]] = :qp2 AND [[name]] = :qp3))', [':qp0' => 1, ':qp1' => 'foo', ':qp2' => 2, ':qp3' => 'bar']],